% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOBCurve.R
\name{OOBCurve}
\alias{OOBCurve}
\title{Out of Bag Learning curve}
\usage{
OOBCurve(mod, measures = list(auc), task, data)
}
\arguments{
\item{mod}{An object of class \code{randomForest} or \code{ranger}, as that created by the 
function \code{\link[randomForest]{randomForest}}/\code{\link[ranger]{ranger}} with option \code{keep.inbag = TRUE}.
Alternatively you can also use a randomForest or ranger model trained with \code{\link[mlr]{train}} of \href{https://github.com/mlr-org/mlr}{mlr}.}

\item{measures}{List of performance measure(s) of mlr to evaluate. Default is auc only.
See the \href{http://mlr-org.github.io/mlr-tutorial/release/html/measures/index.html}{mlr tutorial} for a list of available measures 
for the corresponding task.}

\item{task}{Learning task created by the function \code{\link[mlr]{makeClassifTask}} or \code{\link[mlr]{makeRegrTask}} of \href{https://github.com/mlr-org/mlr}{mlr}.}

\item{data}{Original data that was used for training the random forest.}
}
\value{
Returns a dataframe with a column for each desired measure.
}
\description{
With the help of this function the out of bag learning curve for random forests 
can be created for any measure that is available in the mlr package.
}
\examples{
library(ranger)
library(randomForest)

# Classification
data = getTaskData(sonar.task)
mod = ranger(Class ~., data = data, num.trees = 100, keep.inbag = TRUE)

# Alternatively use randomForest
# mod = randomForest(Class ~., data = data, ntree = 100, keep.inbag = TRUE)
# Alternatively use train of mlr
# mod = train(makeLearner("classif.ranger", keep.inbag = TRUE), sonar.task)

# Application of the main function
results = OOBCurve(mod, measures = list(mmce, auc, brier), task = sonar.task, data = data)
# Plot the generated results
plot(results$mmce, type = "l", ylab = "oob-mmce", xlab = "ntrees")
plot(results$auc, type = "l", ylab = "oob-auc", xlab = "ntrees")
plot(results$brier, type = "l", ylab = "oob-brier-score", xlab = "ntrees")

# Regression
data = getTaskData(bh.task)
mod = ranger(medv ~., data = data, num.trees = 100, keep.inbag = TRUE)
# Application of the main function
results = OOBCurve(mod, measures = list(mse, mae), task = bh.task, data = data)
# Plot the generated results
plot(results$mse, type = "l", ylab = "oob-mse", xlab = "ntrees")
plot(results$mae, type = "l", ylab = "oob-mae", xlab = "ntrees")

}
